requirements.txt
Django==2.0.5
django-ipware==2.1.0
pkg-resources==0.0.0
pytz==2018.4
termcolor==1.1.0
